refactor: format 함수 모듈화
